backend/requirements.txt
flask==3.0.0
PyGithub==2.1.1
pyyaml==6.0.1
gunicorn==21.2.0

# Dev dependencies
black>=24.0.0
mypy>=1.0.0
pytest>=8.0.0
pytest-cov>=4.1.0
